Add Floating Button

+++ app/src/main/java/com/rickyputrah/fooddelivery/ui/home/widget/food/FoodListWidget.kt
@@ -17,7 +17,7 @@ data class FoodListWidgetSpec(
         val name: String,
         val description: String,
         val portion: String,
-        val price: Double
+        val price: Int
     )
 }
